fixed common problems
