+++ src/pages/Dashboard.jsx
@@ -3,15 +3,18 @@ import { useNavigate } from "react-router-dom";
 import { usePageTitle } from "../hooks/usePageTitle";
 import ItemForm from "../components/ItemForm";
 import {
+  deleteExistingItem,
   fetchItems,
 } from "../utils/watchlistHandler";
 import { useLoading } from "../context/loadingContext";
 import useAuthStore, { selectIsLoggedIn } from "../store/useAuthStore";
 import WatchlistTabs from "../components/WatchlistTabs";
+import { useModal } from "../context/modalContext";
 
 export default function Watchlist() {
   usePageTitle("Watchlist");
 
+  const { openModal, closeModal } = useModal();
   const [items, setItems] = useState([]);
   const [isInitialLoad, setIsInitialLoad] = useState(true);
 
@@ -68,6 +71,70 @@ export default function Watchlist() {
     closeModal();
   };
 
+  // ✅ Update existing item
+  const handleItemUpdated = (updatedItem) => {
+    setItems((prevItems) =>
+      prevItems.map((item) => (item.id === updatedItem.id ? updatedItem : item))
+    );
+    closeModal();
+  };
+
+  // ✅ Open modal in update mode
+  const handleUpdateItem = (item) => {
+    openModal(
+      <ItemForm
+        existingItem={item}
+        onItemAdded={handleItemAdded}
+        onItemUpdated={handleItemUpdated}
+      />
+    );
+  };
+
+  // ✅ Delete item with confirmation popup
+  const handleDeleteItem = (id) => {
+    openModal(
+      <div className="p-6 bg-white dark:bg-slate-900 rounded-lg shadow-xl max-w-md w-full text-center border border-transparent transition-all">
+        <h2 className="text-xl font-semibold mb-4 text-gray-900 dark:text-white">
+          Confirm Deletion
+        </h2>
+        <p className="text-gray-600 dark:text-gray-400 mb-6">
+          Are you sure you want to delete this item? This action cannot be
+          undone.
+        </p>
+
+        <div className="flex justify-center gap-4">
+          {/* Cancel Button */}
+          <button
+            className="bg-gray-200 dark:bg-slate-800 text-gray-800 dark:text-gray-200 px-5 py-2 rounded hover:bg-gray-300 dark:hover:bg-slate-700 transition-colors"
+            onClick={closeModal}
+          >
+            Cancel
+          </button>
+
+          {/* Delete Button */}
+          <button
+            className="bg-red-600 text-white px-5 py-2 rounded hover:bg-red-700 dark:hover:bg-red-500 shadow-md shadow-red-500/20 transition-colors"
+            onClick={async () => {
+              try {
+                await deleteExistingItem(id);
+                setItems((prevItems) =>
+                  prevItems.filter((item) => item.id !== id)
+                );
+                closeModal();
+              } catch (error) {
+                console.error("Failed to delete item:", error);
+                closeModal();
+              }
+            }}
+          >
+            Delete
+          </button>
+        </div>
+      </div>,
+      false
+    );
+  };
+
 
   return (
     <div className="min-h-screen p-4 bg-gray-50 dark:bg-slate-950 transition-colors duration-300">
@@ -87,7 +154,14 @@ export default function Watchlist() {
         </button>
       </div>
 
-        <WatchlistTabs items={items} handleItemAdded={handleItemAdded} isInitialLoad={isInitialLoad} reloadItemsFromCloud={reloadItemsFromCloud}  />
+      <WatchlistTabs
+        items={items}
+        handleItemUpdated={handleItemUpdated}
+        handleUpdateItem={handleUpdateItem}
+        handleDeleteItem={handleDeleteItem}
+        isInitialLoad={isInitialLoad}
+        reloadItemsFromCloud={reloadItemsFromCloud}
+      />
 
     </div>
   );
